+++ lib/data/projects/personal/projects-overview/personalNext.ts
@@ -6,11 +6,11 @@ import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 import { Project } from '@/lib/utils/interfaces/interfaces'
 
 // project images
-import krsiak1 from '@/public/images/projects/personal/krsiak/krsiak-1-light.png'
-import krsiak2 from '@/public/images/projects/personal/krsiak/krsiak-2-light.png'
-import krsiak3 from '@/public/images/projects/personal/krsiak/krsiak-3-light.png'
-import krsiak4 from '@/public/images/projects/personal/krsiak/krsiak-4-light.png'
-import krsiak5 from '@/public/images/projects/personal/krsiak/krsiak-5-light.png'
+import krsiak1 from '@/public/images/projects/personal/krsiak/krsiak-1-light.webp'
+import krsiak2 from '@/public/images/projects/personal/krsiak/krsiak-2-light.webp'
+import krsiak3 from '@/public/images/projects/personal/krsiak/krsiak-3-light.webp'
+import krsiak4 from '@/public/images/projects/personal/krsiak/krsiak-4-light.webp'
+import krsiak5 from '@/public/images/projects/personal/krsiak/krsiak-5-light.webp'
 import krsiakThumb from '@/public/images/projects/personal/krsiak/krsiak-thumb-light-600.webp'
 
 export const projectsPersonalNext: Project[] = [
